+++ src/ng2-tabs/ng2-tabs.component.ts
@@ -1,12 +1,20 @@
 import { Component, ContentChildren, AfterContentInit,
   Output, EventEmitter } from '@angular/core'
 import { Ng2TabComponent } from './ng2-tab.component'
+
+interface SelectTabModel {
+  tab?: Ng2TabComponent
+  metaData?: any
+  selectMethod?: string // 'clickedTab'
+  activatorId?: string
+}
+
 @Component({
   selector: 'ng2-tabs',
   template: `
     <ul class="nav nav-tabs">
       <li class="nav-item" *ngFor="let tab of coTabCmps"
-        (click)="selectTab(tab)">
+        (click)="selectTab({tab: tab, selectMethod: 'clickedTab'})">
         <a href="#" class="nav-link"
           [class.active]="tab.active">
           {{tab.title}}
@@ -23,26 +31,44 @@ export class Ng2TabsComponent implements AfterContentInit {
   public ngAfterContentInit () {
     let activeTabs = this.coTabCmps.filter(tab => tab.active)
     if (activeTabs.length === 0) {
-      this.selectTab(this.coTabCmps.first, true)
+      this.selectTab({
+        tab: this.coTabCmps.first,
+        selectMethod: 'initFirstTab'
+      })
     }
   }
 
-  public selectTab (tab, initial?) {
-    this.coTabCmps.toArray().forEach(tab => tab.active = false)
-    tab.active = true
+  public selectTab (options: SelectTabModel) {
+    let foundTab
+    let selectMethod
+
+    // The tab itself
+    if (options.tab) {
+      this.unselectAllTabs()
+      options.tab.active = true
+
+    // Select by activator ID
+    } else if (options.activatorId) {
+      selectMethod = 'activatorId'
+      foundTab = this.coTabCmps.toArray().find(tab => tab.activatorId === options.activatorId)
+      if (!foundTab) {
+        console.error(`No tab with activatorId: "${options.activatorId}" found`)
+      } else {
+        this.unselectAllTabs()
+        foundTab.active = true
+      }
+    }
     this.selectedTab.emit({
-      initial: !!initial,
-      tab
-    })
+      tab: options.tab || foundTab,
+      metaData: options.metaData,
+      selectMethod: options.selectMethod || selectMethod,
+      activatorId: options.activatorId
+    } as SelectTabModel)
+
     return false
   }
 
-  public selectTabByActivatorId (activatorId) {
-    let foundTab = this.coTabCmps.toArray().find(tab => tab.activatorId === activatorId)
-    if (!foundTab) {
-      console.error(`No tab with activatorId: "${activatorId}" found`)
-    } else {
-      this.selectTab(foundTab)
-    }
+  private unselectAllTabs () {
+    this.coTabCmps.toArray().forEach(tab => tab.active = false)
   }
 }