+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0 (2017-01-27)
+
+* selectTabByActivatorId => selectTab({activatorId: 'activatorId'})
+* selectTab() has {metaData} property that can be used to pass through information
+* (selectedTab) now contains more information
+
 ## 1.4.0 (2016-12-14)
 
 * When a tab is selected, ng2-tabs emits "selectedTab(tab)"